# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalyzeRequest": {
        "description": "A single incident signal to assess.",
        "properties": {
          "message": {
            "description": "Raw event message to assess.",
            "maxLength": 8192,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType",
            "default": "unknown",
            "description": "Signal origin. One of: application, infrastructure, network, iam, audit, unknown."
          }
        },
        "required": [
          "message"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
      "AnalyzeResponse": {
        "description": "Structured incident context produced by the Sentinel pipeline.",
        "properties": {
          "analysis_scope": {
            "title": "Analysis Scope",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "integer"
          },
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "title": "Evidence",
            "type": "array"
          },
          "explain": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explain"
          },
          "feedback": {
            "additionalProperties": true,
            "title": "Feedback",
            "type": "object"
          },
          "historical_context": {
            "additionalProperties": true,
            "title": "Historical Context",
            "type": "object"
          },
          "hypothesis": {
            "title": "Hypothesis",
            "type": "string"
          },
          "impact": {
            "additionalProperties": true,
            "title": "Impact",
            "type": "object"
          },
          "impact_feedback_endpoint": {
            "title": "Impact Feedback Endpoint",
            "type": "string"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "monitored_object": {
            "title": "Monitored Object",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "string"
          },
          "recommended_action": {
            "title": "Recommended Action",
            "type": "string"
          },
          "recommended_steps": {
            "items": {
              "type": "string"
            },
            "title": "Recommended Steps",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "risk_score": {
            "title": "Risk Score",
            "type": "number"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "incident_id",
          "analysis_scope",
          "event_type",
          "monitored_object",
          "message",
          "source",
          "risk_score",
          "severity",
          "priority",
          "confidence",
          "recommended_action",
          "evidence",
          "context",
          "historical_context",
          "impact",
          "hypothesis",
          "recommended_steps",
          "feedback",
          "tier",
          "impact_feedback_endpoint"
        ],
        "title": "AnalyzeResponse",
        "type": "object"
      },
      "DecisionImpactRequest": {
        "properties": {
          "action_taken": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Action Taken",
            "type": "string"
          },
          "comments": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "confidence": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Confidence",
            "type": "integer"
          },
          "decision_taken": {
            "enum": [
              "ACCEPT",
              "REVIEW",
              "BLOCK",
              "IGNORE",
              "INVESTIGATE",
              "OTHER"
            ],
            "title": "Decision Taken",
            "type": "string"
          },
          "hypothesis_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis Correct"
          },
          "incident_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Incident Id",
            "type": "string"
          },
          "memory_write_approved": {
            "default": false,
            "title": "Memory Write Approved",
            "type": "boolean"
          },
          "replaced_manual_process": {
            "title": "Replaced Manual Process",
            "type": "boolean"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "resolution_text": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Text"
          },
          "time_saved_minutes": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Time Saved Minutes",
            "type": "integer"
          }
        },
        "required": [
          "incident_id",
          "request_id",
          "decision_taken",
          "action_taken",
          "confidence",
          "replaced_manual_process",
          "time_saved_minutes"
        ],
        "title": "DecisionImpactRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SourceType": {
        "description": "Origin of an ingested signal — drives source-aware risk weighting.",
        "enum": [
          "application",
          "infrastructure",
          "network",
          "iam",
          "audit",
          "unknown"
        ],
        "title": "SourceType",
        "type": "string"
      },
      "ValidateResolutionRequest": {
        "description": "Request body for POST /beta/validate-resolution.\n\nExtends DecisionImpactRequest with optional signal metadata needed to\nconstruct a ValidatedIncident. All new fields are optional so that\nexisting DecisionImpactRequest-compatible payloads remain valid.",
        "properties": {
          "action_taken": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Action Taken",
            "type": "string"
          },
          "comments": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          },
          "confidence": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Confidence",
            "type": "integer"
          },
          "decision_taken": {
            "enum": [
              "ACCEPT",
              "REVIEW",
              "BLOCK",
              "IGNORE",
              "INVESTIGATE",
              "OTHER"
            ],
            "title": "Decision Taken",
            "type": "string"
          },
          "event_type": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Type"
          },
          "hypothesis_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypothesis Correct"
          },
          "incident_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Incident Id",
            "type": "string"
          },
          "memory_write_approved": {
            "default": false,
            "title": "Memory Write Approved",
            "type": "boolean"
          },
          "message_normalized": {
            "anyOf": [
              {
                "maxLength": 8192,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Normalized"
          },
          "replaced_manual_process": {
            "title": "Replaced Manual Process",
            "type": "boolean"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "resolution_text": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Text"
          },
          "signature": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sentinel content signature.",
            "title": "Signature"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "supersedes": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ID of the ValidatedIncident this replaces.",
            "title": "Supersedes"
          },
          "time_saved_minutes": {
            "maximum": 100000.0,
            "minimum": 0.0,
            "title": "Time Saved Minutes",
            "type": "integer"
          },
          "tokens": {
            "items": {
              "type": "string"
            },
            "maxItems": 256,
            "title": "Tokens",
            "type": "array"
          },
          "validated_by": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated By"
          }
        },
        "required": [
          "incident_id",
          "request_id",
          "decision_taken",
          "action_taken",
          "confidence",
          "replaced_manual_process",
          "time_saved_minutes"
        ],
        "title": "ValidateResolutionRequest",
        "type": "object"
      },
      "ValidateResolutionResponse": {
        "description": "Response from POST /beta/validate-resolution.",
        "properties": {
          "decision_impact_id": {
            "title": "Decision Impact Id",
            "type": "string"
          },
          "memory_record_created": {
            "title": "Memory Record Created",
            "type": "boolean"
          },
          "memory_record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Record Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "decision_impact_id",
          "memory_record_created",
          "status"
        ],
        "title": "ValidateResolutionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "## RAVEN Incident Intelligence API\n\nTwo independent scoring engines are available:\n\n| Endpoint | Engine | Score range | Severity bands |\n|---|---|---|---|\n| `POST /evaluate` | Rule-based (legacy) | Integer **0 – 100** | LOW / MEDIUM / HIGH |\n| `POST /v1/analyze` | Sentinel 4-layer pipeline | Float **0.0 – 1.0** | LOW / MEDIUM / HIGH / CRITICAL |\n\nAll protected endpoints require the `X-API-Key` header.\n\nAll error responses share the envelope `{\"error\": \"...\", \"code\": \"...\"}`.",
    "title": "RAVEN Incident Intelligence API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/beta/decision-impact": {
      "post": {
        "description": "Submit after acting on a RAVEN recommendation. Use the `incident_id` and `request_id` from the `/v1/analyze` response.",
        "operationId": "submit_decision_impact_beta_decision_impact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionImpactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Submit Decision Impact Beta Decision Impact Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record that RAVEN influenced a real decision",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/beta/validate-resolution": {
      "post": {
        "description": "Records an analyst's validation of a RAVEN recommendation. Always writes a DecisionImpact record. When `memory_write_approved=true` and `resolution_text` is provided, also creates a ValidatedIncident in long-term memory (`validated_incidents.jsonl`).",
        "operationId": "validate_resolution_beta_validate_resolution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResolutionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit analyst validation and optionally store as incident memory",
        "tags": [
          "feedback"
        ]
      }
    },
...
    "/v1/analyze": {
      "post": {
        "description": "Normalizes and scores one incident signal, then returns structured evidence, an investigation hypothesis, context, and suggested next steps. The result supports human investigation and is not an automated verdict.",
        "operationId": "analyze_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "synthetic_normal_operation": {
                  "summary": "Synthetic normal operation",
                  "value": {
                    "message": "Background job completed successfully",
                    "source": "application"
                  }
                },
                "synthetic_service_failure": {
                  "summary": "Synthetic service failure",
                  "value": {
                    "message": "Checkout API returned HTTP 503 after a configuration change in the synthetic staging environment",
                    "source": "application"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Analyze an incident signal",
        "tags": [
          "v1"
        ]
      }
    }
  }
}